Fix audit bugs: type cast, dashboard alarm sync, settings live data
- Fix uint8_t cast for confidence assignment in simulation_engine
- Add comment clarifying legacy ackAlarm vs new ackDynAlarm in dashboard
- Add Simulation Engine status panel to web dashboard settings page
- Show live sensor values in web settings sensor config section

+++ src/data/simulation_engine.cpp
@@ -621,7 +621,7 @@ void sim_update(void) {
         } else {
             baseConf = 55 + random(0, 25);
         }
-        demo->ai.insights[i].confidence = clampf(baseConf, 40, 99);
+        demo->ai.insights[i].confidence = (uint8_t)clampf((float)baseConf, 40.0f, 99.0f);
     }
 
     // OTA simulation
